more intelligent guesswork, bug fix
Use the first and last character of each word to help predict a word's likelyhood  of being the answer. The more matches, the better.

When putting in '->' or '>' to proceed to cracking the password, the character(s) would end up in the password list. Added a fix to remove them if there.